[Update] Language
Setting 언어전환기능 추가

+++ Invisiphobia_NullPath/Assets/Invisiphobia_NullPath/Scripts/Core/SettingManager.cs
@@ -72,17 +72,17 @@ namespace Common.Setting
             mouseSensitivity = value;
         }
 
-        public static void SetZoomHold(float index)
+        public static void SetZoomHold(int index)
         {
             zoomHold = index == 1 ? true : false;
         }
 
-        public static void SetRunHold(float index)
+        public static void SetRunHold(int index)
         {
             runHold = index == 1 ? true : false;
         }
 
-        public static void SetCrouchHold(float index)
+        public static void SetCrouchHold(int index)
         {
             crouchHold = index == 1 ? true : false;
         }
